# api-spec.yaml
openapi: 3.1.0
info:
  title: MS Template API
  description: "Microservice template API exposing company, salesman, admin user,\
    \ and external posts endpoints. All endpoints support request context propagation\
    \ using X-Request-Id, X-Tenant-Id, and Accept-Language headers."
  contact:
    name: MS Template Team
    url: https://snb.com/support
    email: support@snb.com
  license:
    name: Proprietary
    url: https://snb.com/legal
  version: v1
...
servers:
- url: http://localhost:8080
  description: Local development
- url: /
  description: Relative deployment base path
...
paths:
...
  /api/roles/{roleCode}/privileges:
    get:
      tags:
      - RBAC
      summary: Get privileges by role code
      description: Returns all active privilege grants for a role.
      operationId: getPrivilegesByRoleCode
      parameters:
      - name: roleCode
        in: path
        description: Role code
        required: true
        schema:
          type: string
        example: SUPER_ADMIN
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      responses:
        "200":
          description: Role privileges fetched successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RolePrivilegeResponse"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
    post:
      tags:
      - RBAC
      summary: Grant privilege to role
      description: Creates a role-privilege grant.
      operationId: grantPrivilegeToRole
      parameters:
      - name: roleCode
        in: path
        description: Role code
        required: true
        schema:
          type: string
        example: SUPER_ADMIN
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      requestBody:
        description: Grant payload
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RolePrivilegeRequest"
            examples:
              GrantPrivilege:
                description: GrantPrivilege
                value:
                  privilegeCode: USER_VIEW
        required: true
      responses:
        "201":
          description: Privilege granted successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RolePrivilegeResponse"
        "400":
          description: Validation failed or privilege already granted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "404":
          description: Role or privilege not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
  /api/roles/{roleCode}/privileges/bulk:
    post:
      tags:
      - RBAC
      summary: Grant privileges to role in bulk
      description: Creates multiple role-privilege grants in one request.
      operationId: grantPrivilegesToRoleBulk
      parameters:
      - name: roleCode
        in: path
        description: Role code
        required: true
        schema:
          type: string
        example: SUPER_ADMIN
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      requestBody:
        description: Bulk grant payload
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RolePrivilegeBulkRequest"
            examples:
              GrantPrivilegesBulk:
                description: GrantPrivilegesBulk
                value:
                  privilegeCodes:
                  - USER_VIEW
                  - USER_EDIT
        required: true
      responses:
        "201":
          description: Privileges granted successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RolePrivilegeResponse"
        "400":
          description: Validation failed or privilege already granted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "404":
          description: Role or privilege not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
  /api/roles/{roleCode}/privileges/{privilegeCode}:
    delete:
      tags:
      - RBAC
      summary: Revoke role-privilege grant
      description: Soft-deletes a role-privilege grant by roleCode and privilegeCode.
      operationId: revokeRolePrivilege
      parameters:
      - name: roleCode
        in: path
        description: Role code
        required: true
        schema:
          type: string
        example: SUPER_ADMIN
      - name: privilegeCode
        in: path
        description: Privilege code
        required: true
        schema:
          type: string
        example: USER_VIEW
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      responses:
        "200":
          description: Privilege revoked successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RolePrivilegeResponse"
        "404":
          description: Grant not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
...
components:
  schemas:
...
    RolePrivilegeRequest:
      type: object
      description: Role-privilege grant request payload
      properties:
        privilegeCode:
          type: string
          description: Privilege code to grant
          example: USER_VIEW
      required:
      - privilegeCode
    RolePrivilegeBulkRequest:
      type: object
      description: Role-privilege bulk grant request payload
      properties:
        privilegeCodes:
          type: array
          description: Privilege codes to grant
          items:
            type: string
            example: USER_VIEW
      required:
      - privilegeCodes
    RolePrivilegeResponse:
      type: object
      description: Role-privilege grant response payload
      properties:
        id:
          type: integer
          format: int64
          description: Grant record identifier
          example: 1
        roleId:
          type: integer
          format: int64
          description: Role identifier
          example: 1
        roleCode:
          type: string
          description: Role code
          example: SUPER_ADMIN
        privilegeId:
          type: integer
          format: int64
          description: Privilege identifier
          example: 1
        privilegeCode:
          type: string
          description: Privilege code
          example: USER_VIEW
        privilegeName:
          type: string
          description: Privilege name
          example: View Users
        createdBy:
          type: integer
          format: int64
          description: Identifier of actor who created the record
          example: 1
        createdAt:
          type: string
          format: date-time
          description: Record creation timestamp
          example: 2026-05-19T10:00:00
        updatedBy:
          type: integer
          format: int64
          description: Identifier of actor who last updated the record
          example: 1
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: 2026-05-19T11:00:00
...
    ErrorInfo:
      type: object
      description: Detailed error information
      properties:
        type:
          type: string
          description: Error category
          example: VALIDATION_ERROR
        code:
          type: string
          description: Application-specific error code
          example: EMAILADDRESS_INVALID
        message:
          type: string
          description: "Localized, user-friendly error message"
          example: emailAddress must be a valid email
        description:
          type: string
          description: Additional diagnostic details
          example: "Rejected value: not-an-email"
    BaseResponseDTO:
      type: object
      description: Base API response wrapper with error details
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorInfo"
...
  parameters:
    XRequestIdHeader:
      name: X-Request-Id
      in: header
      description: "Correlation identifier for distributed tracing. If omitted, the\
        \ service generates one."
      required: false
      schema:
        type: string
        example: req-20260504-0001
    XTenantIdHeader:
      name: X-Tenant-Id
      in: header
      description: Tenant identifier used for multi-tenant context propagation.
      required: false
      schema:
        type: string
        example: tenant-001
    AcceptLanguageHeader:
      name: Accept-Language
      in: header
      description: Preferred locale for localized validation and error messages.
      required: false
      schema:
        type: string
        example: en